fix commit

+++ src/main/java/faang/school/urlshortenerservice/encoder/Base62.java
@@ -6,11 +6,11 @@ import org.springframework.stereotype.Component;
 public class Base62 {
     static final char[] BASE62 = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789".toCharArray();
 
-    public static String base62(int n) {
+    public static String base62(Long n) {
         final StringBuilder sb = new StringBuilder();
 
         while (n > 0) {
-            sb.append(BASE62[n % 62]);
+            sb.append(BASE62[(int) (n % 62)]);
             n = n / 62;
         }
         return sb.reverse().toString();
