+++ src/main/java/faang/school/urlshortenerservice/encoder/Base62Encoder.java
@@ -5,5 +5,5 @@ import faang.school.urlshortenerservice.model.Hash;
 import java.util.List;
 
 public interface Base62Encoder {
-    List<Hash> encode(List<Integer> numbers);
+    List<Hash> encode(List<Long> numbers);
 }
